fix(DTFS2-6829): refactor facility tenor to be more dry based on review comments
