+++ e2e-tests/tfm/cypress/e2e/journeys/case-amendments/user-flow/7. amendment-details-automatic-approval.spec.js
@@ -1,38 +1,15 @@
-import { add } from 'date-fns';
 import relative from '../../../relativeURL';
 import { errorSummary } from '../../../partials';
 import facilityPage from '../../../pages/facilityPage';
 import amendmentsPage from '../../../pages/amendments/amendmentsPage';
 import caseDealPage from '../../../pages/caseDealPage';
 import MOCK_DEAL_AIN from '../../../../fixtures/deal-AIN';
-import { fourDaysAgo, oneMonth, today, twoMonths, yearWithZeroLetter, twoYearsAgo } from '../../../../../../e2e-fixtures/dateConstants';
+import { fourDaysAgo, oneMonth, today, twoMonths, yearWithZeroLetter } from '../../../../../../e2e-fixtures/dateConstants';
 import { PIM_USER_1, UNDERWRITER_MANAGER_DECISIONS, BANK1_MAKER1, ADMIN, CURRENCY } from '../../../../../../e2e-fixtures';
+import { FACILITY_TENOR } from '../../../../fixtures/facility-tenor';
 
 context('Amendments - automatic approval journey', () => {
-  /**
-   * The deal in the test is a BSS deal.
-   *
-   * The tenor is calculated using a SQL function used by MDM.
-   *
-   * In these tests, the mock start date is two years ago from today
-   * & the mock end date is a month from today.
-   *
-   * So in this case it boils down to the tenor being 25 months except for:
-   * - if expiry and commencement dates are the same date of month then you add one
-   * - if commencement is end of month and expiry is also end of month you add one
-   */
-  let facilityTenor = '25 months';
-
-  if (twoYearsAgo.date.getDate() === oneMonth.date.getDate()) {
-    facilityTenor = '26 months';
-  }
-
-  const isCommencementDateEndOfMonth = add(twoYearsAgo.date, { days: 1 }).getDate() === 1;
-  const isExpiryDateEndOFMonth = add(oneMonth.date, { days: 1 }).getDate() === 1;
-
-  if (isCommencementDateEndOfMonth && isExpiryDateEndOFMonth) {
-    facilityTenor = '26 months';
-  }
+  const facilityTenor = FACILITY_TENOR.BSS.COMMENCEMENT_TWO_YEARS_AGO_EXPIRY_IN_ONE_MONTH;
 
   describe('Amendment details - Change the Cover end date AND Facility value', () => {
     let dealId;
